Days 4-6.
On day 6, I entered a wrong answer for the first star. The bug was that
I captured the set of points that are disqualified (by having an
infinite area) but then wasn't using that set to actually rule out those
regions; the number I entered was the area of one of those infinite
regions that fell into the clipping rectangle I happened to be using.

Dafny of course would have caught this error, under any remotely
reasonable specification.