2019 day 2 part 2, finished.
Of course there's a much easier way, but I did the fun thing. :)

This was too complicated and too much of a conversation with the machine
to check in a "first" version; the first version I ran was not going to
work. Bugs I had include:

-   Using an int where I needed a Number. This led to nodes like
    `Add([2, Number(2)])` which rendered, mysteriously, as `2 + 2`.

-   Using a Number where I needed an int. This led to nodes like
    `Number(Number(1))` that rendered as `1`.

Both of these would have been caught, trivially, by a type system.
(Or type-checking assertions in constructors--the same thing but
horrible.)

-   Saying `add` when I meant `mul`. (This would not have been caught,
    and is hard to debug when you have insufficient tests.)

Not bugs, exactly, but false starts:

-   All of the knowledge that is now in `simplify_add` and
    `simplify_mul` was previously in `__add__` and `__mul__` methods.
    That is, the code was more object-oriented. But this was a
    failure. Cases kept slipping through without being simplified,
    and the choice was between centralization and duplication.
    I went with the former.

-   Previously `Add(a, b)` and `Mul(a, b)` were simple binary
    constructors that did not have a separate `.constant` field.
    This was hostile to simplification.