2019 day 12 part 2.
I was really sleepy when I wrote part 1, on the edge of falling asleep
at my computer, so the code is some of the more scattered I've written
this year.

This puzzle would have been fun to do in Rust, because it's a
simulation; the speed would be nice, and mutability with immutability
guarantees when you want them is great. I used deepcopy in the Python
code! I also changed the Vec3 type to be a namedtuple. It is slow.

I had to fix a lot of minor bugs to get this working, all trivial as far
as I can remember.