Finished part 1.
I was feeling drowsy and stupid today, and the program is kind of
redundant and ugly. It lacks any internal boundaries; in typical OO
fashion, I started by designing a Robot class, and it slowly gained
responsibilities until it contained all the state and was doing all the
work.

Lots of typos and forgot-to-import errors; an API misuse error; a bug
subclassing `tuple`, something I will probably never understand
properly; forgot to add a default argument in `LabeledDigraph.__init__`;
a typo-like bug (using `pop` instead of `popleft`) which made
`shortest_path` do a depth-first search rather than a breadth-first
search; a bug where `Robot.explore()` checked whether a new node is
visited *after* adding it to the data structure, so it always seems
visited; and, the bug leading to me submitting my first wrong answer of
the year: the algorithm measured the distance from the starting position
to the robot's final position (`robot.location`), rather than to the
oxygen system (`robot.target`). Off by one!